requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
parameterized==0.9.0
requests==2.32.4
urllib3==2.5.0
djangorestframework-simplejwt
